# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PlanOut": {
        "properties": {
          "business_code": {
            "title": "Business Code",
            "type": "string"
          },
          "config_version": {
            "title": "Config Version",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "integer"
          },
          "default_resource_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Default Resource Ids",
            "type": "array"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_enabled": {
            "title": "Is Enabled",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "business_code",
          "description",
          "default_resource_ids",
          "config_version",
          "is_enabled",
          "created_by",
          "created_at"
        ],
        "title": "PlanOut",
        "type": "object"
      },
      "PlanWrite": {
        "properties": {
          "business_code": {
            "enum": [
              "fut_mm",
              "rem_two",
              "rem_two_mm"
            ],
            "title": "Business Code",
            "type": "string"
          },
          "config_version": {
            "default": "1.0",
            "title": "Config Version",
            "type": "string"
          },
          "default_resource_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Default Resource Ids",
            "type": "array"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "is_enabled": {
            "default": true,
            "title": "Is Enabled",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "business_code"
        ],
        "title": "PlanWrite",
        "type": "object"
      },
...
      "ScenarioOut": {
        "properties": {
          "config_version": {
            "title": "Config Version",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "default_resource_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Default Resource Ids",
            "type": "array"
          },
          "draft_workflow_version_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Draft Workflow Version Id"
          },
          "expected_artifacts": {
            "items": {
              "type": "string"
            },
            "title": "Expected Artifacts",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_archived": {
            "title": "Is Archived",
            "type": "boolean"
          },
          "is_enabled": {
            "title": "Is Enabled",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "plan_id": {
            "title": "Plan Id",
            "type": "integer"
          },
          "published_workflow_version_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published Workflow Version Id"
          },
          "required_resource_types": {
            "items": {
              "type": "string"
            },
            "title": "Required Resource Types",
            "type": "array"
          },
          "scenario_type": {
            "title": "Scenario Type",
            "type": "string"
          },
          "workflow_status": {
            "title": "Workflow Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "plan_id",
          "name",
          "scenario_type",
          "config_version",
          "expected_artifacts",
          "default_resource_ids",
          "required_resource_types",
          "is_enabled",
          "workflow_status",
          "draft_workflow_version_id",
          "published_workflow_version_id",
          "is_archived",
          "created_at"
        ],
        "title": "ScenarioOut",
        "type": "object"
      },
      "ScenarioWrite": {
        "additionalProperties": false,
        "properties": {
          "config_version": {
            "default": "1.0",
            "title": "Config Version",
            "type": "string"
          },
          "default_resource_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default Resource Ids"
          },
          "expected_artifacts": {
            "items": {
              "type": "string"
            },
            "title": "Expected Artifacts",
            "type": "array"
          },
          "is_enabled": {
            "default": true,
            "title": "Is Enabled",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "plan_id": {
            "title": "Plan Id",
            "type": "integer"
          },
          "required_resource_types": {
            "items": {
              "type": "string"
            },
            "title": "Required Resource Types",
            "type": "array"
          },
          "scenario_type": {
            "title": "Scenario Type",
            "type": "string"
          }
        },
        "required": [
          "plan_id",
          "name",
          "scenario_type"
        ],
        "title": "ScenarioWrite",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "OpenSLT",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/plans": {
      "get": {
        "operationId": "list_plans_api_v1_plans_get",
        "parameters": [
          {
            "in": "query",
            "name": "business_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Business Code"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanOut"
                  },
                  "title": "Response List Plans Api V1 Plans Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Plans"
      },
      "post": {
        "operationId": "create_plan_api_v1_plans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Plan"
      }
    },
    "/api/v1/plans/{plan_id}": {
      "delete": {
        "operationId": "delete_plan_api_v1_plans__plan_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Plan"
      },
      "put": {
        "operationId": "update_plan_api_v1_plans__plan_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Plan"
      }
    },
    "/api/v1/plans/{plan_id}/copy": {
      "post": {
        "operationId": "copy_plan_api_v1_plans__plan_id__copy_post",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Copy Plan"
      }
    },
...
    "/api/v1/scenarios": {
      "get": {
        "operationId": "list_scenarios_api_v1_scenarios_get",
        "parameters": [
          {
            "in": "query",
            "name": "plan_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plan Id"
            }
          },
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ScenarioOut"
                  },
                  "title": "Response List Scenarios Api V1 Scenarios Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Scenarios"
      },
      "post": {
        "operationId": "create_scenario_api_v1_scenarios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Scenario"
      }
    },
    "/api/v1/scenarios/{scenario_id}": {
      "delete": {
        "operationId": "delete_scenario_api_v1_scenarios__scenario_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Scenario"
      },
      "put": {
        "operationId": "update_scenario_api_v1_scenarios__scenario_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Scenario"
      }
    },
    "/api/v1/scenarios/{scenario_id}/copy": {
      "post": {
        "operationId": "copy_scenario_api_v1_scenarios__scenario_id__copy_post",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Copy Scenario"
      }
    },
...
  }
}